\name{oldbooks}
\docType{data}
\alias{oldbooks}
\title{A collection of five classic books}
\description{
   This object is a list containing five classical books:
   \itemize{
   \item \emph{Around the World in Eighty Days} by Jules Verne
   \item \emph{The Three Musketeers} by Alexandre Dumas
   \item \emph{Frankenstein} by Mary Shelley
   \item \emph{Dracula} by Bram Stoker
   \item \emph{The Strange Case of Dr Jekyll and Mr Hyde} by Robert Stevenson}
   
   as single-element character vectors. All five books were taken from the \href{http://www.gutenberg.org}{Project Gutenberg homepage} and contain formatting symbols, such as \\n for breaks.
   }
\usage{data(oldbooks)}
\format{A named list containing five character vectors as elements}
\source{\href{http://www.gutenberg.org}{Project Gutenberg}}
\references{ 
 Dumas, A. (1844). \emph{The Three Musketeers}. Retrieved from\cr http://www.gutenberg.org/ebooks/1257
 
 Shelley, M. W. (1818). \emph{Frankenstein; Or, The Modern Prometheus}. Retrieved from\cr http://www.gutenberg.org/ebooks/84
 
 Stevenson, R. L. (1886). \emph{The Strange Case of Dr. Jekyll and Mr. Hyde}. Retrieved from\cr http://www.gutenberg.org/ebooks/42
 
 Stoker, B. (1897). \emph{Dracula}. Retrieved from\cr http://www.gutenberg.org/ebooks/345
 
 Verne, J.(1873). \emph{Around the World in Eighty Days}. Retrieved from\cr http://www.gutenberg.org/ebooks/103
 
 

}
\keyword{Books}