\name{priming}
\docType{data}
\alias{priming}
\title{Simulated data for a Semantic Priming Experiment}
\description{
   A data frame containing simulated data for a Semantic Priming Experiment. This data contains 514 prime-target pairs, which are taken from the Hutchison, Balota, Cortese and Watson (2008) study. These pairs are generated by pairing each of 257 target words with one semantically related and one semantically unrelated prime.\cr
   The data frame contains four columns:
   \itemize{
   \item First column: Prime Words
   \item Second column: Target Words
   \item Third column: \strong{Simulated} Reaction Times
   \item Fourth column: Specifies whether a prime-target pair is considered semantically related or unrelated
   }
   }
\usage{data(priming)}
\format{A data frame with 514 rows and 4 columns}

\references{
Hutchison, K. A., Balota, D. A., Cortese, M. & Watson, J. M. (2008). Predicting semantic priming at the item level. \emph{Quarterly Journal of Experimental Psychology, 61,} 1036-1066.
}

\keyword{Synonym Test}