\name{syntest}
\docType{data}
\alias{syntest}
\title{A multiple choice test for synonyms and antonyms}
\description{
   This object multiple choice test for synonyms and antonyms, consisting of seven columns. 
   \enumerate{
   \item The first column defines the question, i.e. the word a synonym or an antonym has to be found for. 
  \item The second up to the fifth column show the possible answer alternatives.
   \item The sixth column defines the correct answer.
   \item The seventh column indicates whether a synonym or an antonym has to be found for the word in question.
   }
   The test consists of twenty questions, which are given in the twenty rows of the data frame.
   }
\usage{data(syntest)}
\format{A data frame with 20 rows and 7 columns}

\keyword{Synonym Test}